COMMERCE-1489  Code improvement and unused code removal
