+++ modules/apps/commerce/commerce-product-content-search-web/src/main/resources/com/liferay/commerce/product/content/cpspecificationoption/facet/template/dependencies/portlet_display_template_label.ftl
@@ -32,13 +32,5 @@
 				</#list>
 			</div>
 		</#if>
-
-		<#if !cpSpecificationOptionsSearchFacetDisplayContext.isNothingSelected()>
-			<@liferay_aui.button
-				cssClass="btn-link btn-unstyled facet-clear-btn"
-				onClick="Liferay.Search.FacetUtil.clearSelections(event);"
-				value="clear"
-			/>
-		</#if>
 	</@>
 </@>
